requirements.txt
yt-dlp
notebooklm-py[browser]
playwright
pytest
ruff
duckduckgo-search
arxiv
